+++ src/Server/Configuration/WikiBlazorServerOptions.cs
@@ -1,6 +1,7 @@
 ﻿using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.DependencyInjection;
 using Tavenem.Wiki.Blazor.Client;
+using Tavenem.Wiki.Blazor.Client.Configuration;
 using Tavenem.Wiki.Blazor.Server.Authorization;
 using Tavenem.Wiki.Blazor.Services;
 
@@ -9,7 +10,7 @@ namespace Tavenem.Wiki.Blazor.Server.Configuration;
 /// <summary>
 /// Options for configuring <c>Tavenem.Wiki.Blazor.Server</c>.
 /// </summary>
-public class WikiBlazorServerOptions() : WikiBlazorOptions
+public class WikiBlazorServerOptions() : WikiBlazorClientOptions
 {
     private IFileManager? _fileManager;
     private Func<IServiceProvider, IFileManager>? _fileManagerConfig;
@@ -113,11 +114,6 @@ public class WikiBlazorServerOptions() : WikiBlazorOptions
         AboutPageTitle = other.AboutPageTitle;
         AppBar = other.AppBar;
         AppBarRenderMode = other.AppBarRenderMode;
-        ArticleEndMatter = other.ArticleEndMatter;
-        ArticleEndMatterRenderMode = other.ArticleEndMatterRenderMode;
-        ArticleFrontMatter = other.ArticleFrontMatter;
-        ArticleFrontMatterRenderMode = other.ArticleFrontMatterRenderMode;
-        CanEditOffline = other.CanEditOffline;
         CategoriesTitle = other.CategoriesTitle;
         CategoryNamespace = other.CategoryNamespace;
         CompactLayout = other.CompactLayout;
@@ -135,21 +131,15 @@ public class WikiBlazorServerOptions() : WikiBlazorOptions
         DefaultTableOfContentsTitle = other.DefaultTableOfContentsTitle;
         DomainArchivePermission = other.DomainArchivePermission;
         FileNamespace = other.FileNamespace;
-        GetDomainPermission = other.GetDomainPermission;
         GroupNamespace = other.GroupNamespace;
         HelpPageTitle = other.HelpPageTitle;
         InteractiveRenderMode = other.InteractiveRenderMode;
-        IsOfflineDomain = other.IsOfflineDomain;
         LinkTemplate = other.LinkTemplate;
         LoginPath = other.LoginPath;
         MainLayout = other.MainLayout;
         MainPageTitle = other.MainPageTitle;
         MaxFileSize = other.MaxFileSize;
         MinimumTableOfContentsHeadings = other.MinimumTableOfContentsHeadings;
-        OnCreated = other.OnCreated;
-        OnDeleted = other.OnDeleted;
-        OnEdited = other.OnEdited;
-        OnRenamed = other.OnRenamed;
         PolicyPageTitle = other.PolicyPageTitle;
         Postprocessors = other.Postprocessors;
         ScriptNamespace = other.ScriptNamespace;
@@ -161,6 +151,18 @@ public class WikiBlazorServerOptions() : WikiBlazorOptions
         UserNamespace = other.UserNamespace;
         WikiLinkPrefix = other.WikiLinkPrefix;
         WikiServerApiRoute = other.WikiServerApiRoute;
+    }
+
+    /// <summary>
+    /// Constructs a new instance of <see cref="WikiBlazorServerOptions"/>.
+    /// </summary>
+    /// <param name="other">
+    /// An instance of <see cref="WikiBlazorClientOptions"/> from which to copy settings.
+    /// </param>
+    public WikiBlazorServerOptions(WikiBlazorClientOptions other) : this(other as WikiBlazorOptions)
+    {
+        ArticleRenderManager = other.ArticleRenderManager;
+        OfflineManager = other.OfflineManager;
     }
 
     /// <summary>